+++ app/src/main/java/woowacourse/shopping/domain/model/OrderSummary.kt
@@ -1,10 +1,10 @@
 package woowacourse.shopping.domain.model
 
 data class OrderSummary(
-    val orderAmount: Int,
+    val orderAmount: Long,
     val couponAmount: Int,
     val shippingFee: Int,
 ) {
-    val totalPrice: Int
-        get() = orderAmount - couponAmount + shippingFee
+    val totalPrice: Long
+        get() = (orderAmount - couponAmount + shippingFee)
 }
